视频转gif功能的一些调整2
