+++ demo/src/main/java/cn/jzvd/demo/CustomJzvd/JzvdStdGif.java
@@ -1,11 +1,15 @@
 package cn.jzvd.demo.CustomJzvd;
 
+import java.io.File;
+
 import android.content.Context;
 import android.util.AttributeSet;
 import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
-import android.widget.SeekBar;
+import android.widget.FrameLayout;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import cn.jzvd.Jzvd;
 import cn.jzvd.JzvdStd;
@@ -19,6 +23,9 @@ public class JzvdStdGif extends JzvdStd {
 
     GifCreateHelper mGifCreateHelper;
 
+    TextView tv_hint;
+    FrameLayout fl_hint_region;
+
     public JzvdStdGif(Context context) {
         super(context);
     }
@@ -30,27 +37,47 @@ public class JzvdStdGif extends JzvdStd {
     @Override
     public void init(Context context) {
         super.init(context);
+
+        tv_hint = findViewById(R.id.tv_hint);
+        fl_hint_region = findViewById(R.id.fl_hint_region);
+
+        initGifHelper();
+        findViewById(R.id.convert_to_gif).setOnClickListener((v -> {
+            clickStartGif(v);
+        }));
+
+        topContainer.setOnTouchListener(new OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return true;
+            }
+        });
+        fl_hint_region.setOnTouchListener(new OnTouchListener() {
+            @Override
+            public boolean onTouch(View v, MotionEvent event) {
+                return true;
+            }
+        });
     }
 
     /**
-     *
-     * @param jaoziVideoGifSaveListener  保存进度的回调函数
+     * @param jaoziVideoGifSaveListener 保存进度的回调函数
      */
     public void initGifHelper(GifCreateHelper.JaoziVideoGifSaveListener jaoziVideoGifSaveListener) {
         mGifCreateHelper = new GifCreateHelper(this, jaoziVideoGifSaveListener);
     }
 
     /**
-     * @param jaoziVideoGifSaveListener  保存进度的回调函数
-     * @param delay          每一帧之间的延时
-     * @param inSampleSize   采样率，最小值1 即：每隔inSampleSize个像素点，取一个读入到内存。越大处理越快
-     * @param smallScale     缩小倍数，越大处理越快
-     * @param gifPeriod      gif时长，毫秒
-     * @param gifPath        gif文件的存储路径
+     * @param jaoziVideoGifSaveListener 保存进度的回调函数
+     * @param delay                     每一帧之间的延时
+     * @param inSampleSize              采样率，最小值1 即：每隔inSampleSize个像素点，取一个读入到内存。越大处理越快
+     * @param smallScale                缩小倍数，越大处理越快
+     * @param gifPeriod                 gif时长，毫秒
+     * @param gifPath                   gif文件的存储路径
      */
     public void initGifHelper(GifCreateHelper.JaoziVideoGifSaveListener jaoziVideoGifSaveListener,
-                               int delay, int inSampleSize, int smallScale, int gifPeriod, String gifPath) {
-        mGifCreateHelper = new GifCreateHelper(this,jaoziVideoGifSaveListener,delay,inSampleSize,smallScale,gifPeriod,gifPath);
+                              int delay, int inSampleSize, int smallScale, int gifPeriod, String gifPath) {
+        mGifCreateHelper = new GifCreateHelper(this, jaoziVideoGifSaveListener, delay, inSampleSize, smallScale, gifPeriod, gifPath);
     }
 
     /**
@@ -58,7 +85,7 @@ public class JzvdStdGif extends JzvdStd {
      */
     public void startGif() {
         mGifCreateHelper.startGif();
-        if(Jzvd.STATE_PLAYING==state){
+        if (Jzvd.STATE_PLAYING == state) {
             startButton.performClick();
         }
 
@@ -66,169 +93,67 @@ public class JzvdStdGif extends JzvdStd {
 
     /**
      * 开始生成gif
-     * @param bitmapFromTime  gif图在视频中的开始时间
-     * @param vedioUrl         视频链接
+     *
+     * @param bitmapFromTime gif图在视频中的开始时间
+     * @param vedioUrl       视频链接
      */
-    public void startGif(long bitmapFromTime,String vedioUrl) {
-        mGifCreateHelper.startGif(bitmapFromTime,vedioUrl);
-        if(Jzvd.STATE_PLAYING==state){
+    public void startGif(long bitmapFromTime, String vedioUrl) {
+        mGifCreateHelper.startGif(bitmapFromTime, vedioUrl);
+        if (Jzvd.STATE_PLAYING == state) {
             startButton.performClick();
         }
     }
 
     @Override
-    public void onClick(View v) {
-        super.onClick(v);
-        int i = v.getId();
-        if (i == cn.jzvd.R.id.fullscreen) {
-            Log.i(TAG, "onClick: fullscreen button");
-        } else if (i == R.id.start) {
-            Log.i(TAG, "onClick: start button");
-        }
-    }
-
-    @Override
-    public boolean onTouch(View v, MotionEvent event) {
-        super.onTouch(v, event);
-        int id = v.getId();
-        if (id == cn.jzvd.R.id.surface_container) {
-            switch (event.getAction()) {
-                case MotionEvent.ACTION_UP:
-                    if (mChangePosition) {
-                        Log.i(TAG, "Touch screen seek position");
-                    }
-                    if (mChangeVolume) {
-                        Log.i(TAG, "Touch screen change volume");
-                    }
-                    break;
-            }
-        }
-
-        return false;
-    }
-
-    @Override
     public int getLayoutId() {
-        return R.layout.jz_layout_std;
+        return R.layout.jz_layout_gif;
     }
 
-    @Override
-    public void startVideo() {
-        super.startVideo();
-        Log.i(TAG, "startVideo");
+    long current;
+    private void initGifHelper() {
+        //详细使用方式
+//        myJzvdStd.initGifHelper(jaoziVideoGifSaveListener,delay,inSampleSize,smallScale,gifPeriod, gifPath)
+
+        //快速使用方式
+        initGifHelper(new GifCreateHelper.JaoziVideoGifSaveListener() {
+            @Override
+            public void result(boolean success, File file) {
+                post(new Runnable() {
+                    @Override
+                    public void run() {
+                        fl_hint_region.setVisibility(View.GONE);
+                        Toast.makeText(getContext(), "创建成功", Toast.LENGTH_LONG).show();
+                    }
+                });
+            }
+
+            @Override
+            public void process(int curPosition, int total,String status) {
+                Log.e("fffs",status+"  "+curPosition+"/"+total+"  curruentTime: "+(System.currentTimeMillis()-current));
+                post(new Runnable() {
+                    @Override
+                    public void run() {
+                        tv_hint.setText(status+"  "+curPosition+"/"+total);
+                    }
+                });
+            }
+        });
     }
 
-    @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        super.onStopTrackingTouch(seekBar);
-        Log.i(TAG, "Seek position ");
-    }
+    public void clickStartGif(View view) {
+        tv_hint.setText("正在创建Gif...");
+        fl_hint_region.setVisibility(View.VISIBLE);
+        current=System.currentTimeMillis();
+        startGif();
 
-    @Override
-    public void gotoScreenFullscreen() {
-        super.gotoScreenFullscreen();
-        Log.i(TAG, "goto Fullscreen");
+//        HttpProxyCacheServer proxy = ApplicationDemo.getProxy(this);
+//        //startGif的两种使用方式
+//        if(proxy.isCached(urlCrude)){
+//            //被缓存了直接读取缓存
+//            startGif();
+//        }else{
+//            //没被缓存读取原始链接（速度较慢）
+//            startGif(getCurrentPositionWhenPlaying(),urlCrude);
+//        }
     }
-
-    @Override
-    public void gotoScreenNormal() {
-        super.gotoScreenNormal();
-        Log.i(TAG, "quit Fullscreen");
-    }
-
-    @Override
-    public void autoFullscreen(float x) {
-        super.autoFullscreen(x);
-        Log.i(TAG, "auto Fullscreen");
-    }
-
-    @Override
-    public void onClickUiToggle() {
-        super.onClickUiToggle();
-        Log.i(TAG, "click blank");
-    }
-
-    //onState 代表了播放器引擎的回调，播放视频各个过程的状态的回调
-    @Override
-    public void onStateNormal() {
-        super.onStateNormal();
-    }
-
-    @Override
-    public void onStatePreparing() {
-        super.onStatePreparing();
-    }
-
-    @Override
-    public void onStatePlaying() {
-        super.onStatePlaying();
-    }
-
-    @Override
-    public void onStatePause() {
-        super.onStatePause();
-    }
-
-    @Override
-    public void onStateError() {
-        super.onStateError();
-    }
-
-    @Override
-    public void onStateAutoComplete() {
-        super.onStateAutoComplete();
-        Log.i(TAG, "Auto complete");
-    }
-
-    //changeUiTo 真能能修改ui的方法
-    @Override
-    public void changeUiToNormal() {
-        super.changeUiToNormal();
-    }
-
-    @Override
-    public void changeUiToPreparing() {
-        super.changeUiToPreparing();
-    }
-
-    @Override
-    public void changeUiToPlayingShow() {
-        super.changeUiToPlayingShow();
-    }
-
-    @Override
-    public void changeUiToPlayingClear() {
-        super.changeUiToPlayingClear();
-    }
-
-    @Override
-    public void changeUiToPauseShow() {
-        super.changeUiToPauseShow();
-    }
-
-    @Override
-    public void changeUiToPauseClear() {
-        super.changeUiToPauseClear();
-    }
-
-    @Override
-    public void changeUiToComplete() {
-        super.changeUiToComplete();
-    }
-
-    @Override
-    public void changeUiToError() {
-        super.changeUiToError();
-    }
-
-    @Override
-    public void onInfo(int what, int extra) {
-        super.onInfo(what, extra);
-    }
-
-    @Override
-    public void onError(int what, int extra) {
-        super.onError(what, extra);
-    }
-
 }
