+++ demo/src/main/res/layout/activity_gif.xml
@@ -14,48 +14,17 @@
             android:layout_height="wrap_content"
             android:orientation="vertical">
 
-            <FrameLayout
+            <LinearLayout
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="200dp">
+                android:layout_height="200dp">
 
-                    <cn.jzvd.demo.CustomJzvd.JzvdStdGif
-                        android:id="@+id/jz_video"
-                        android:layout_width="match_parent"
-                        android:layout_height="200dp" />
-                </LinearLayout>
-                
-                <ImageView
-                    android:id="@+id/convert_to_gif"
-                    android:layout_width="50dp"
-                    android:layout_height="wrap_content"
-                    android:adjustViewBounds="true"
-                    android:padding="10dp"
-                    android:layout_gravity="right|center_vertical"
-                    android:layout_marginRight="3dp"
-                    android:layout_marginTop="40dp"
-                    android:src="@mipmap/ic_record_start"/>
-            </FrameLayout>
+                <cn.jzvd.demo.CustomJzvd.JzvdStdGif
+                    android:id="@+id/jz_video"
+                    android:layout_width="match_parent"
+                    android:layout_height="200dp" />
+            </LinearLayout>
             
 
         </LinearLayout>
     </ScrollView>
-
-    <FrameLayout
-        android:id="@+id/fl_hint_region"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#88000000"
-        android:visibility="gone">
-
-        <TextView
-            android:id="@+id/tv_hint"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="#ffffff"
-            android:layout_gravity="center"
-            android:text="正在创建Gif..."/>
-    </FrameLayout>
 </FrameLayout>
